Create HRT_MNT_P3
Here are proposed changes for the code of the heart rate monitor, the old code could only function in the second version of Processing. The proposed changes now work for the newest version (3.3.2). I figured other individuals could benefit from this and help them get their project done faster and with a lot less headaches.

**Side-Note**
The only hiccup I encountered in modifying the code and comparing it to the old one's readings was that when it is supposed to be flat-line there is a negative charge that is registered, which was not there when run in Processing v2.